// dmifonFE/src/main/webapp/app/models/proEntcon.model.ts
export interface IproEntcon {
  id: number;
  idcap: string;
  idcro: string;
  impass: number;
  impeco: number;
  impliq: number;
  impman: number;
  impmaneme: number;
  nroconapp: number;
  tipent: string;
  anncmp: number;
  codcig: string;
  codcup: string;
  codentcon: string;
  codgami: string;
  desimp: string;
  despdd: string;
}